+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -41,18 +41,23 @@ import type { ExtractEventPayload } from '@metamask/messenger';
 import type { SnapId } from '@metamask/snaps-sdk';
 import { type Snap } from '@metamask/snaps-utils';
 import { assert, mask, object, string } from '@metamask/superstruct';
-import type { Hex, Json, SemVerVersion } from '@metamask/utils';
+import type {
+  Hex,
+  Json,
+  SemVerVersion,
+  DeferredPromise,
+} from '@metamask/utils';
 import {
   bigIntToHex,
   hasProperty,
   KnownCaipNamespace,
   toCaipChainId,
+  createDeferredPromise,
 } from '@metamask/utils';
 import { Mutex } from 'async-mutex';
 import { v4 as uuid } from 'uuid';
 
 import { transformAccount } from './account';
-import { DeferredPromise } from './DeferredPromise';
 import {
   AccountCreatedEventStruct,
   AccountUpdatedEventStruct,
@@ -494,7 +499,7 @@ export class SnapKeyring {
     // letting the MetaMask client that this "save" has been run.
     // NOTE: Another way of fixing that could be to rely on events through the
     // messenger maybe?
-    const onceSaved = new DeferredPromise<AccountId>();
+    const onceSaved = createDeferredPromise<AccountId>();
 
     // Add the account to the keyring, but wait for the MetaMask client to
     // approve the account creation first.
@@ -1334,7 +1339,7 @@ export class SnapKeyring {
     requestId: string,
     snapId: SnapId,
   ): DeferredPromise<Response> {
-    const promise = new DeferredPromise<Response>();
+    const promise = createDeferredPromise<Response>();
     this.#requests.set(requestId, { promise, snapId });
     return promise;
   }